move sleep to timer.sleep
