--- src/async/sleep.test.ts
@@ -1,17 +1,0 @@
-import { sleep } from './sleep'
-test('should sleep work', async () => {
-  let result = 0
-  async function testSleep() {
-    await sleep(50)
-    result = 1
-  }
-  expect(result).toBe(0)
-  testSleep().then(() => {
-    expect(result).toBe(1)
-  })
-  expect(result).toBe(0)
-  await sleep(40)
-  expect(result).toBe(0)
-  await sleep(55)
-  expect(result).toBe(1)
-})
