--- src/async/sleep.ts
@@ -1,7 +1,0 @@
-/**
- * Sleep millisecond
- * @param ms
- */
-export function sleep(ms: number) {
-  return new Promise(res => setTimeout(res, ms))
-}
